CA-203817: Remove /usr/lib64/ocaml/xapi-storage from base package
Change 5db0cd was intended to move all OCaml files out of the base
xapi-storage package, so that installing it would no longer pull
in gcc.   That change did not remove %{_libdir}/ocaml/xapi-storage
from the base package, so the OCaml files were still pulled into
it, causing it to depend on ocaml and gcc.

Signed-off-by: Euan Harris <euan.harris@citrix.com>